refactor: move phrases code to /app/phrases
